feat features/string: add setup search options and improve search
New function to setup search options interactively or programmatically.

It also improves the search function to use these options.
The search function now asks for file extension if it's not provided.
The search options include smart case, case sensitive and ignore case.
These options can be toggled on and off interactively using
the new key mapping '<Leader>fs'. The search function also handles the
case when the search key is not provided.