plugin: cmp: update nvim-lspconfig compatible
require('lspconfig') (the legacy "framework" of nvim-lspconfig) is deprecated in favor of vim.lsp.config (Nvim 0.11+).
The lspconfig.lua module will be dropped. Calls to require('lspconfig') will show a warning, which will later become an error.
nvim-lspconfig itself is NOT deprecated. It provides server-specific configs.
The configs live in the lsp/ directory. vim.lsp.config automatically finds them and merges them with any local lsp/*.lua configs defined by you or a plugin.
The old configs in lua/lspconfig/ are deprecated and will be removed.
Migration instructions

Upgrade to Nvim 0.11+
(Optional) Use vim.lsp.config('…') (not require'lspconfig'.….setup{}) to customize or define a config.
Use vim.lsp.enable('…') (not require'lspconfig'.….setup{}) to enable a config, so that it activates for its filetypes.